Add files via upload
HTML  and  Javascript:Create a front end that allows the user to enter details of the item purchased (item name, item price (option), number of items bought). On clicking of the submit button, the total cost for the item purchased should be calculated and displayed. On clicking on the ‘paid’ button, appropriate alert box should be popped up